# clinica_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clínica Social Cuidar - API System",
    "description": "\n**API Oficial de Gestão da Clínica Social Cuidar.**\n\n## Funcionalidades Principais:\n*   **Gestão de Pacientes**: Cadastro completo com criptografia de dados sensíveis (LGPD).\n*   **Prontuário Eletrônico**: Registro histórico de atendimentos.\n*   **Controle Financeiro**: Receitas e Despesas.\n*   **Auditoria**: Rastreabilidade total de ações.\n\n## Segurança e Governança:\n*   Todos os endpoints são protegidos (JWT).\n*   Rate Limiting ativo para prevenção de ataques.\n*   Dados sensíveis (CPF) são criptografados no banco (Fernet).\n        ",
    "contact": {
      "name": "Suporte Técnico - Vitor Moraes",
      "email": "suporte@clinicasocial.com.br"
    },
    "license": {
      "name": "Private License"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/financial/transactions": {
      "post": {
        "tags": [
          "financial"
        ],
        "summary": "Create Transaction",
        "operationId": "create_transaction_api_v1_financial_transactions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "financial"
        ],
        "summary": "Get Transactions",
        "operationId": "get_transactions_api_v1_financial_transactions_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "date_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date Filter"
            }
          },
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "name": "appointment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response Get Transactions Api V1 Financial Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/financial/transactions/detail/{transaction_id}": {
      "get": {
        "tags": [
          "financial"
        ],
        "summary": "Get Transaction",
        "operationId": "get_transaction_api_v1_financial_transactions_detail__transaction_id__get",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/financial/stats/daily": {
      "get": {
        "tags": [
          "financial"
        ],
        "summary": "Get Daily Stats",
        "operationId": "get_daily_stats_api_v1_financial_stats_daily_get",
        "parameters": [
          {
            "name": "target_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Target Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/financial/stats/summary": {
      "get": {
        "tags": [
          "financial"
        ],
        "summary": "Get Financial Summary",
        "operationId": "get_financial_summary_api_v1_financial_stats_summary_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/financial/transactions/{id}": {
      "delete": {
        "tags": [
          "financial"
        ],
        "summary": "Delete Transaction",
        "operationId": "delete_transaction_api_v1_financial_transactions__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "financial"
        ],
        "summary": "Update Transaction",
        "operationId": "update_transaction_api_v1_financial_transactions__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailyStats": {
        "properties": {
          "total_income": {
            "type": "number",
            "title": "Total Income"
          },
          "total_expense": {
            "type": "number",
            "title": "Total Expense"
          },
          "balance": {
            "type": "number",
            "title": "Balance"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count"
          },
          "by_method": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "By Method"
          }
        },
        "type": "object",
        "required": [
          "total_income",
          "total_expense",
          "balance",
          "transaction_count",
          "by_method"
        ],
        "title": "DailyStats"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "DINHEIRO",
          "PIX",
          "CARTAO",
          "OUTRO"
        ],
        "title": "PaymentMethod"
      },
...
      "Transaction": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "appointment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod",
            "default": "DINHEIRO"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "type",
          "description"
        ],
        "title": "Transaction"
      },
      "TransactionCreate": {
        "properties": {
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "appointment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "type",
          "description",
          "date",
          "payment_method"
        ],
        "title": "TransactionCreate"
      },
      "TransactionResponse": {
        "properties": {
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "appointment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "type",
          "description",
          "date",
          "payment_method",
          "id",
          "created_at"
        ],
        "title": "TransactionResponse"
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "INCOME",
          "EXPENSE"
        ],
        "title": "TransactionType"
      },
      "TransactionUpdate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "appointment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "payment_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "TransactionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}